+++ buffa-test/src/lib.rs
@@ -100,6 +100,16 @@ pub mod json_types {
 #[allow(
     clippy::derivable_impls,
     clippy::match_single_binding,
+    non_camel_case_types,
+    dead_code
+)]
+pub mod view_json {
+    buffa::include_proto!("test.viewjson");
+}
+
+#[allow(
+    clippy::derivable_impls,
+    clippy::match_single_binding,
     non_camel_case_types
 )]
 pub mod p2json {